Clean up, refactoring, dep update
