+++ module/core/collection_tools/Cargo.toml
@@ -34,7 +34,6 @@ no_std = [
 ]
 use_alloc = [
   "no_std",
-  "hashbrown",
   "test_tools/use_alloc",
 ]
 
@@ -51,16 +50,15 @@ full = [
 enabled = []
 
 # Collection constructors, like `hmap!{ "key" => "val" }`
-collection_constructors = [ "literally" ]
+collection_constructors = []
 # STD collection for no_std.
-collection_std = [ "hashbrown" ]
+collection_std = []
 
 
 [dependencies]
 
 ## external
-literally = { version = "~0.1.3", optional = true, default-features = false }
-hashbrown = { version = "~0.14.3", optional = true, default-features = false, features = [ "default" ] }
+hashbrown = { version = "~0.14.3", default-features = false, features = [ "default" ] }
 
 [dev-dependencies]
 test_tools = { workspace = true }
